Move BrowserLayout to its own file
