+++ qt/aqt/profiles.py
@@ -10,7 +10,7 @@ import shutil
 import traceback
 from enum import Enum
 from pathlib import Path
-from typing import Any
+from typing import TYPE_CHECKING, Any
 
 import anki.lang
 import aqt.forms
@@ -21,10 +21,13 @@ from anki.lang import without_unicode_isolation
 from anki.sync import SyncAuth
 from anki.utils import int_time, is_mac, is_win, point_version
 from aqt import appHelpSite
-from aqt.enums import BrowserLayout
 from aqt.qt import *
 from aqt.theme import Theme
 from aqt.utils import disable_help_button, send_to_trash, showWarning, tr
+
+if TYPE_CHECKING:
+    from aqt.browser.layout import BrowserLayout
+
 
 # Profile handling
 ##########################################################################
@@ -544,6 +547,8 @@ create table if not exists profiles
         self.meta["theme"] = theme.value
 
     def browser_layout(self) -> BrowserLayout:
+        from aqt.browser.layout import BrowserLayout
+
         return BrowserLayout(self.meta.get("browser_layout", "auto"))
 
     def set_browser_layout(self, layout: BrowserLayout) -> None:
